chore: 아이디 값 int 타입에서 bigint(Long)타입으로 변경

+++ src/main/java/com/ohgiraffers/warehousemanagement/wms/inspection/service/InspectionServiceImpl.java
@@ -62,7 +62,7 @@ public class InspectionServiceImpl implements InspectionService {
             throw new IllegalArgumentException("존재하지 않는 유형입니다!");
     }
 
-    public InspectionResponseDTO getInspectionById(int inspectionId) {
+    public InspectionResponseDTO getInspectionById(Long inspectionId) {
         Optional<Inspection> findInspection = inspectionRepository.findById(inspectionId);
         if (findInspection.isPresent()) {
             return new InspectionResponseDTO(findInspection.get());
@@ -90,7 +90,7 @@ public class InspectionServiceImpl implements InspectionService {
     }
 
     @Transactional
-    public InspectionResponseDTO updateInspection(int inspectionId, InspectionRequestDTO requestDTO) {
+    public InspectionResponseDTO updateInspection(Long inspectionId, InspectionRequestDTO requestDTO) {
         Optional<Inspection> findInspection = inspectionRepository.findById(inspectionId);
         if (findInspection.isPresent()) {
             Inspection inspection = findInspection.get();
